% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODDPub.R
\name{open_data_search}
\alias{open_data_search}
\title{Seach for open data & open code keywords.}
\usage{
open_data_search(PDF_text_sentences)
}
\arguments{
\item{PDF_text_sentences}{Document corpus loaded with the pdf_load function.}
}
\value{
Tibble with one row per screened document and the filename and logical values for open data
and open code detection as columns.
}
\description{
The algorithm searches for several categories of similar keywords in each sentence.
Multiple categories have to match for a single sentence to trigger a detection.
Among keyword categories are categories for specific biomedical databases as well as
their corresponding accession numbers (as regular expressions), general-purpose repositories
or different file formats typically used to distribute raw data in the supplement.
Additionally, Open Code dissemination is detected using keywords categories for source code or code repositories.
}
\examples{
open_data_search(pdf_load("examples/"))

}
