% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODDPub.R
\name{open_data_search_parallel}
\alias{open_data_search_parallel}
\title{Parallelizied of the open data & open code keyword search}
\usage{
open_data_search_parallel(
  PDF_text_sentences,
  detected_sentences = TRUE,
  cluster_num = 4
)
}
\arguments{
\item{PDF_text_sentences}{Document corpus loaded with the pdf_load function.}

\item{detected_sentences}{Logical parameter. If TRUE, the sentences in which the Open Data
statements were detected are added to the results table as well.}

\item{cluster_num}{Number of parallel processes that are started. The keyword search is parallelized
with respect to the documents to speed up the calculations.}
}
\value{
Tibble with one row per screened document and the filename and logical values for open data
and open code detection as columns plus additional columns containing the identified open data/code categories
as well as the sentences in which open data/code was detected.
}
\description{
The algorithm searches for several categories of similar keywords in each sentence.
Multiple categories have to match for a single sentence to trigger a detection.
Among keyword categories are categories for specific biomedical databases as well as
their corresponding accession numbers (as regular expressions), general-purpose repositories
or different file formats typically used to distribute raw data in the supplement.
Additionally, Open Code dissemination is detected using keywords categories for source code or code repositories.
}
\examples{
open_data_search_parallel(pdf_load("examples/"))

}
