% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODDPub.R
\name{open_data_sentences}
\alias{open_data_sentences}
\title{Seach for open data & open code keywords and return detected sentences.}
\usage{
open_data_sentences(PDF_text_sentences)
}
\arguments{
\item{PDF_text_sentences}{Document corpus loaded with the pdf_load function.}
}
\value{
Tibble with one row per screened document and the filename and logical values for open data
and open code detection as columns plus an additional column containing the detected sentences
for each of the checked keyword category.
}
\description{
Does the same keyword search as the function \code{\link{open_data_search}}
but also returns the sentences in which the Open Data keywords were detected.
}
\examples{
open_data_sentences(pdf_load("examples/"))
}
