% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODDPub.R
\name{pdf_convert}
\alias{pdf_convert}
\title{Conversion of PDFs to text.}
\usage{
pdf_convert(
  pdf_folder,
  output_folder,
  recursive = TRUE,
  overwrite_existing_files = FALSE,
  add_section_tags = TRUE
)
}
\arguments{
\item{pdf_folder}{String of the folder name in which the PDFs are located.}

\item{output_folder}{String of the folder name in which the converted files will be saved.}

\item{recursive}{Boolean. If TRUE (the default), then search all the subfolders of the given folder for PDF files.}

\item{overwrite_existing_files}{Boolean. If FALSE (the default) does not overwrite already existing files.}

\item{add_section_tags}{Boolean. If TRUE (the default), adds '<section>' tags at the beginning of potential text sections.
This needs to be set to TRUE for later recognition of Data and Code Availability Statements.}
}
\value{
Logical vector describing the conversion success for each PDF file.
}
\description{
Converts PDFs to text before the text can be loaded and text-mined.
For the PDFconversion requires the pdftools library.
for parallelisation requires the future and furrr libraries.
}
