% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODDPub.R
\name{splice_margin_text}
\alias{splice_margin_text}
\title{Splice sentences from margin text (important text in left margin in e.g. PLoS journals)
across two pages and put all margin text first.}
\usage{
splice_margin_text(text_sentences)
}
\arguments{
\item{text_sentences}{Vector of text sentences, typically from a
document corpus loaded with the pdf_load function.}
}
\value{
Data frame of the text with metadata about text in margins.
}
\description{
Splice sentences from margin text (important text in left margin in e.g. PLoS journals)
across two pages and put all margin text first.
}
