% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_text.R
\name{splice_plos_twopager}
\alias{splice_plos_twopager}
\title{Splice sentences from e.g. DAS that were separated due to interpolated
 sections in PLoS-formatted articles.}
\usage{
splice_plos_twopager(sections_v)
}
\arguments{
\item{sections_v}{Character vector of sentences that may contain
interpolated sentences from abstract or other sections.}
}
\value{
Character vector with sentences with interpolated sentences
removed.
}
\description{
Splice sentences from e.g. DAS that were separated due to interpolated
 sections in PLoS-formatted articles.
}
