% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODDPub.R
\name{pdf_convert}
\alias{pdf_convert}
\title{Conversion of PDFs to text.}
\usage{
pdf_convert(PDF_folder, output_folder, recursive = TRUE)
}
\arguments{
\item{PDF_folder}{String of the folder name in which the PDFs are located.}

\item{output_folder}{String of the folder name in which the converted files will be saved.}

\item{recursive}{Boolean. If TRUE (the default), then search all the subfolders of the given folder for PDF files.}
}
\value{
Logical vector describing the conversion success for each PDF file.
}
\description{
Converts PDFs to text before the text can be loaded and text-mined.
For the PDFconversion requires the pdftools library.
for parallelisation requires the future and furrr libraries.
}
